fix: emote upload page chat
